profile: extract usePostExposeState hook to eliminate duplication
Consolidates the expose reference path, isExposed, and publicPostUrl
computations from four separate components into a single reusable hook.
